backend/requirements.txt
opencv-python==4.12.0.88
flask==3.1.2
flask_cors==6.0.1
google-genai==1.46.0
gunicorn==23.0.0
imutils==0.5.4
numpy<2.3.0
scipy>=1.10.0
pydantic==2.12.3
pytesseract==0.3.13
python-dotenv==1.0.1
supabase